interpreter: move some of CustomTarget's args to type_checking
As there are so many wrappers that need these as well
